requirements.txt
alembic==1.11.1
anyio==3.7.0
arrow==1.2.3
click==8.1.3
emoji==2.5.1
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Levenshtein==0.21.1
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.25.0
pandas==2.0.2
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2023.3
rapidfuzz==3.1.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
thefuzz==0.19.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
boto3==1.28.12
boto3-stubs==1.28.15.post1
mypy-boto3-s3==1.28.12
pytest==7.4.0
httpx==0.24.1
PyJWT==2.8.0
lunr==0.6.2
bcrypt==4.0.1
cryptography==41.0.3